refactor(control): remove duplicated code
